Network sharing via multicast
Working network sharing of changes via multicast UDP.
Report username on network shared changes to differentiate between users making
changes.
netio was passing on it's own messages (that it sent to the multicast group)
back to it's listeners. This was incorrect and a more robust check is now made
to avoid this case.
NOTE: This does not account for 2 daemons running the same host.

clipoller reports previous sha on changed branches