CLI options to control logging to socket and file, and log level at CLI
Log output can be configured to go to a file or a socket. The two can be used at
the same time. The level is also configurable. By default the socket is set to
the system default syslog endpoint. This leaves stdout free to report seen
changes.
Note: OSX seems to make it impossible to see anything below Error on syslog.