gitsync/netio uses peer source IP and port to unique change origins and avoid echoes
Uniquing the source of a change is useful when routing or presenting it. We
would use the source IP (GitChange.HostIP) for this but that could be aliased if
mulitple instances of a daemon ran on the same machine (or seemed to for some
reason). Furthermore, gitsync.NetIO would use the user to protect itself from
echoes of changes it sent to the multicast group. This breaks when the same user
runs on multiple machines.
In all cases, it is better to unique for the network layer using the IP:Port
pair (since this has to be unique to deliver correctly). The addition of
GitChange.PeerAddr also indicates if change has come from the network or not.