gitsyncd uses a combined bus to share all GitChanges
We previously wired channels between components directly, limited by the fact
that multiple readers on a channel competed for messages. This was fine but
restricted the combinations of components we could compose without complicated
channel wiring. Switching to a shared bus, where a message published is
delivered to all listeners at most once, allows each module discretion on how to
react. It does force some filtering logic to exist in modules that read and
write to this bus, such as gitsync.NetIO but the check is trivial.
The implementation has not changed functionality but does allow refactoring to
allow more modules. An example would be separate modules to handle local git
hook management and another to present a web API. Both are provided both local
and remote GitChanges and can respond as appropriate.